add khalti and solve image delete problem for defalut img

+++ controller/admin/product/productController.js
@@ -3,11 +3,13 @@ const Product = require("../../../model/productModel")
 
 exports.createProduct = async (req, res) => {
   const file = req.file
-  let filePath
+  let filePath;
+  let defaultImage = true;
   if (!file) {
     filePath = "https://upload.wikimedia.org/wikipedia/commons/thumb/a/a1/Momo_nepal.jpg/1280px-Momo_nepal.jpg"
   } else {
     filePath = file.filename
+    defaultImage = false;
   }
 
   const { productName, productDescription, productPrice, productStatus, productStockQty } = req.body
@@ -24,7 +26,7 @@ exports.createProduct = async (req, res) => {
     productPrice,
     productStatus,
     productStockQty,
-    productImage: "http://localhost:8000/" + filePath
+    productImage: defaultImage ? filePath : process.env.BACKEND_URL + filePath
   })
 
   res.status(200).json({
@@ -59,10 +61,12 @@ exports.deleteProduct = async (req, res) => {
 
   const pathAfterCut = deleteProduct.productImage.slice(lengthToCut);
 
-  fs.unlinkSync('./uploads' + pathAfterCut, err => {
-    if (err) console.log("Error while file deletion", err);
-    else console.log("File deleted sucessfully")
-  })
+  if (deleteProduct.productImage.startsWith(process.env.BACKEND_URL)) {
+    fs.unlinkSync('./uploads' + pathAfterCut, err => {
+      if (err) console.log("Error while file deletion", err);
+      else console.log("File deleted sucessfully")
+    })
+  }
 
   res.status(200).json({
     status: "Success",
@@ -75,8 +79,6 @@ exports.editProduct = async (req, res) => {
   const { id } = req.params;
 
   const { productName, productDescription, productPrice, productStatus, productStockQty } = req.body
-
-  console.log(productDescription, productName, productPrice, productStatus, productStockQty)
 
   if (!productName || !productDescription || !productPrice || !productStatus || !productStockQty) {
     return res.status(400).json({
@@ -96,14 +98,14 @@ exports.editProduct = async (req, res) => {
 
   const pathAfterCut = oldData.productImage.slice(lengthToCut);
 
-  console.log(pathAfterCut)
 
-  if (req.file && req.file.filename) {
+  if (req.file && req.file.filename && oldData.productImage.startsWith(process.env.BACKEND_URL)) {
     fs.unlinkSync('./uploads' + pathAfterCut, err => {
       if (err) console.log("Error while file deletion", err);
       else console.log("File deleted sucessfully")
     })
   }
+
   const updatedData = await Product.findByIdAndUpdate(id, {
     productName,
     productDescription,
